Fix dead code: Attach fileInputRef in DawnMark and centralized file handling
- Moved file handling logic (state, upload, cleanup) from `UploadPanel` to `DawnMark`.
- Attached the previously unused `fileInputRef` to a new hidden `<input>` in `DawnMark`.
- Created `components/types.ts` for shared `BlobEntry` interface.
- Created `components/file-utils.ts` for shared helper functions.
- Enabled file upload from both Editor panel and Uploads panel by sharing the handler.
- Verified with manual Playwright script and visual check.
