+++ components/types.ts
@@ -1,0 +1,6 @@
+export interface BlobEntry {
+  id: string;
+  file: File;
+  url: string;
+  snippet: string;
+}
